Corrected file references after moving example into new micro path
