+++ tensorflow/lite/micro/examples/micro_mnist/model/mnist_model.cc
