+++ tensorflow/lite/micro/examples/micro_mnist/model/mnist_train_and_export_to_tfl_micro.py
