+++ tensorflow/lite/micro/examples/micro_mnist/model/progress_bar.py
